Complete multisheller cli tool
The existing version of the multisheller tool just saved some files in
the /tmp directory. This commit complete support for the tool.

By default, the converted scripts are saved in the same directory
of the input file, just changing the extension to the extension of the
script. The user can specify a different output with the --output option.

The tool now automatically creates a <output>.sh script that automatically
calls the appropriate script (either <output>.bash or <output>.zsh) based
on wheter the script is running under bash or zsh.
See https://github.com/mamba-org/multisheller/issues/11 for more details.

The README is extended with a basic example on the use of the cli tool.

Fix https://github.com/mamba-org/multisheller/issues/15
Fix https://github.com/mamba-org/multisheller/issues/11